Add csv path to spring boot arguments
Add the csv path to the Spring application startup. This will work by
requiring the user to select a csv file path with the following format
when starting a Spring application in the terminal:

mvn spring-boot:run -Dspring-boot.run.arguments="--csvPath=users.csv"

This is done inside `SpringBatchExample1Application`, the main class for
the application.