Add invalid lines and average salary to logs
Here I have added invalid lines with the following format:

Number of errors: NN
Lines with errors:
emp_id,firstName,lastName,dep_id,salary_date,salary.
...

and also the calculation of the average salary per employee.

The display of invalid rows is done in a hacky way and therefore needs
to be refactored.